requirements.txt
absl-py                      2.1.0
annotated-types              0.7.0
anyio                        4.4.0
astunparse                   1.6.3
cachetools                   5.3.3
certifi                      2024.2.2
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
distlib                      0.3.8
dnspython                    2.6.1
email_validator              2.1.1
exceptiongroup               1.2.1
fastapi                      0.111.0
fastapi-cli                  0.0.4
filelock                     3.14.0
flatbuffers                  24.3.25
gast                         0.4.0
google-auth                  2.29.0
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.64.0
h11                          0.14.0
h5py                         3.11.0
httpcore                     1.0.5
httptools                    0.6.1
httpx                        0.27.0
idna                         3.7
jax                          0.4.28
Jinja2                       3.1.4
keras                        2.12.0
libclang                     18.1.1
Markdown                     3.6
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
mdurl                        0.1.2
ml-dtypes                    0.3.2
namex                        0.0.8
numpy                        1.23.5
oauthlib                     3.2.2
opt-einsum                   3.3.0
optree                       0.11.0
orjson                       3.10.3
packaging                    24.0
pillow                       10.3.0
pip                          24.0
platformdirs                 4.2.2
protobuf                     4.25.3
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.7.2
pydantic_core                2.18.3
Pygments                     2.18.0
python-dotenv                1.0.1
python-multipart             0.0.9
PyYAML                       6.0.1
requests                     2.32.3
requests-oauthlib            2.0.0
rich                         13.7.1
rsa                          4.9
scipy                        1.13.1
setuptools                   65.5.0
shellingham                  1.5.4
six                          1.16.0
sniffio                      1.3.1
starlette                    0.37.2
tensorboard                  2.12.3
tensorboard-data-server      0.7.2
tensorflow                   2.12.0
tensorflow-estimator         2.12.0
tensorflow-hub               0.16.1
tensorflow-intel             2.12.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.4.0
tf-keras                     2.15.0
typer                        0.12.3
typing_extensions            4.12.0
ujson                        5.10.0
urllib3                      2.2.1
uvicorn                      0.30.0
virtualenv                   20.26.2
watchfiles                   0.22.0
websockets                   12.0
Werkzeug                     3.0.3
wheel                        0.43.0
wrapt                        1.14.1